Extract shared exercise card from live workout
